Feature: Add initial bundles functionality.
The bundles of items are predefined in a json file that is loaded if there aren't any bundles saved locally.
Bundles are suggested in the same way as items from the catalog. Selecting a bundle adds the items of the bundle to the list.

Future work will allow bundles to be defined on the device.